Fixes + Clear Screen
There were 'or' statements where 'and' statements were suppose to be, so that broke a few things. Added clear screen feature (made it large so it works in fullscreen too).  We still need the transfer feature, but we'll do that Thursday together. along with brainstorming the time-out idea, and anything else we want to do with this.